chore(proofs): use core pex implementation
