+++ packages/core/package.json
@@ -21,7 +21,6 @@
     "prepublishOnly": "yarn run build"
   },
   "dependencies": {
-    "@aries-framework/presentation-exchange": "^0.4.2",
     "@digitalcredentials/jsonld": "^5.2.1",
     "@digitalcredentials/jsonld-signatures": "^9.3.1",
     "@digitalcredentials/vc": "^1.1.2",
