+++ packages/core/src/modules/proofs/services/PresentationExchangeService.ts
@@ -6,6 +6,7 @@ import type { InputDescriptorToCredentials, PresentationSubmission } from '../mo
 import type {
   IPresentationDefinition,
   PresentationSignCallBackParams,
+  Validated,
   VerifiablePresentationResult,
 } from '@sphereon/pex'
 import type {
@@ -13,15 +14,16 @@ import type {
   PresentationSubmission as PexPresentationSubmission,
   PresentationDefinitionV1,
 } from '@sphereon/pex-models'
-import type { OriginalVerifiableCredential } from '@sphereon/ssi-types'
+import type { IVerifiablePresentation, OriginalVerifiableCredential } from '@sphereon/ssi-types'
 
-import { PEVersion, PEX, PresentationSubmissionLocation } from '@sphereon/pex'
+import { Status, PEVersion, PEX, PresentationSubmissionLocation } from '@sphereon/pex'
 import { injectable } from 'tsyringe'
 
 import { getJwkFromKey } from '../../../crypto'
 import { AriesFrameworkError } from '../../../error'
 import { JsonTransformer } from '../../../utils'
 import { getKeyFromVerificationMethod, DidsApi } from '../../dids'
+import { PresentationExchangeError } from '../../presentation-exchange'
 import { SignatureSuiteRegistry, W3cPresentation, W3cCredentialService, ClaimFormat } from '../../vc'
 import { W3cCredentialRepository } from '../../vc/repository'
 import { selectCredentialsForRequest } from '../utils/credentialSelection'
@@ -32,7 +34,9 @@ import {
 } from '../utils/transform'
 
 export type ProofStructure = Record<string, Record<string, Array<W3cVerifiableCredential>>>
-export type PresentationDefinition = IPresentationDefinition
+export type PresentationDefinition = IPresentationDefinition & Record<string, unknown>
+
+export type VerifiablePresentation = IVerifiablePresentation & Record<string, unknown>
 
 @injectable()
 export class PresentationExchangeService {
@@ -49,6 +53,50 @@ export class PresentationExchangeService {
     const holderDids = didRecords.map((didRecord) => didRecord.did)
 
     return selectCredentialsForRequest(presentationDefinition, credentialRecords, holderDids)
+  }
+
+  public validatePresentationDefinition(presentationDefinition: PresentationDefinition) {
+    const validation = PEX.validateDefinition(presentationDefinition)
+    const errorMessages = this.formatValidated(validation)
+    if (errorMessages.length > 0) {
+      throw new PresentationExchangeError(
+        `Invalid presentation definition. The following errors were found: ${errorMessages.join(', ')}`
+      )
+    }
+  }
+
+  public validatePresentationSubmission(presentationSubmission: PexPresentationSubmission) {
+    const validation = PEX.validateSubmission(presentationSubmission)
+    const errorMessages = this.formatValidated(validation)
+    if (errorMessages.length > 0) {
+      throw new PresentationExchangeError(
+        `Invalid presentation submission. The following errors were found: ${errorMessages.join(', ')}`
+      )
+    }
+  }
+
+  public validatePresentation(presentationDefinition: PresentationDefinition, presentation: VerifiablePresentation) {
+    const { errors } = this.pex.evaluatePresentation(presentationDefinition, presentation)
+
+    if (errors) {
+      const errorMessages = this.formatValidated(errors as Validated)
+      if (errorMessages.length > 0) {
+        throw new PresentationExchangeError(
+          `Invalid presentation. The following errors were found: ${errorMessages.join(', ')}`
+        )
+      }
+    }
+  }
+
+  private formatValidated(v: Validated) {
+    return Array.isArray(v)
+      ? (v
+          .filter((r) => r.tag === Status.ERROR)
+          .map((r) => r.message)
+          .filter((m) => Boolean(m)) as Array<string>)
+      : v.tag === Status.ERROR && typeof v.message === 'string'
+      ? [v.message]
+      : []
   }
 
   /**
