--- packages/presentation-exchange/jest.config.ts
@@ -1,14 +1,0 @@
-import type { Config } from '@jest/types'
-
-import base from '../../jest.config.base'
-
-import packageJson from './package.json'
-
-process.env.TZ = 'GMT'
-
-const config: Config.InitialOptions = {
-  ...base,
-  displayName: packageJson.name,
-}
-
-export default config
