--- packages/presentation-exchange/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@aries-framework/presentation-exchange",
-  "main": "build/index",
-  "types": "build/index",
-  "version": "0.4.2",
-  "files": [
-    "build"
-  ],
-  "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "homepage": "https://github.com/openwallet-foundation/agent-framework-javascript/tree/main/packages/presentation-exchange",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/openwallet-foundation/agent-framework-javascript",
-    "directory": "packages/presentation-exchange"
-  },
-  "scripts": {
-    "build": "yarn run clean && yarn run compile",
-    "clean": "rimraf ./build",
-    "compile": "tsc -p tsconfig.build.json",
-    "prepublishOnly": "yarn run build"
-  },
-  "dependencies": {
-    "@aries-framework/core": "^0.4.2",
-    "@sphereon/pex": "^2.2.2",
-    "@sphereon/pex-models": "^2.1.2",
-    "@sphereon/ssi-types": "^0.17.5",
-    "jsonpath": "^1.1.1",
-    "tsyringe": "^4.8.0"
-  },
-  "devDependencies": {
-    "@types/jsonpath": "^0.2.4",
-    "typescript": "~4.9.5"
-  }
-}
