--- packages/presentation-exchange/src/PresentationExchangeError.ts
@@ -1,3 +1,0 @@
-import { AriesFrameworkError } from '@aries-framework/core'
-
-export class PresentationExchangeError extends AriesFrameworkError {}
