--- packages/presentation-exchange/src/index.ts
@@ -1,4 +1,0 @@
-export * from './PresentationExchangeError'
-export * from './PresentationExchangeModule'
-export * from './PresentationExchangeService'
-export * from './models'
