--- packages/presentation-exchange/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './transform'
-export * from './credentialSelection'
