--- packages/presentation-exchange/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-
-  "compilerOptions": {
-    "outDir": "./build"
-  },
-
-  "include": ["src/**/*", "types"]
-}
