--- packages/presentation-exchange/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "allowJs": false,
-    "typeRoots": ["../../node_modules/@types", "src/types"],
-    "types": ["jest"]
-  }
-}
